Rename example Xcode project
