+++ Examples/HelloWorldiOSClientAppExample/HelloWorldiOSClientApp.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
