+++ Examples/HelloWorldiOSClientAppExample/HelloWorldiOSClientApp.xctestplan
@@ -4,30 +4,30 @@
       "id" : "89964AC6-C28B-4C2A-9CF4-824B569065D0",
       "name" : "Test Scheme Action",
       "options" : {
-
+        "targetForVariableExpansion" : null
       }
     }
   ],
   "defaultOptions" : {
     "targetForVariableExpansion" : {
-      "containerPath" : "container:iOSAppClient.xcodeproj",
+      "containerPath" : "container:HelloWorldiOSClientApp.xcodeproj",
       "identifier" : "3A133CDE2B1F28870008DD5E",
-      "name" : "iOSAppClient"
+      "name" : "HelloWorldiOSClientApp"
     }
   },
   "testTargets" : [
     {
       "target" : {
-        "containerPath" : "container:iOSAppClient.xcodeproj",
+        "containerPath" : "container:HelloWorldiOSClientApp.xcodeproj",
         "identifier" : "3A133CEE2B1F28890008DD5E",
-        "name" : "iOSAppClientTests"
+        "name" : "HelloWorldiOSClientAppTests"
       }
     },
     {
       "target" : {
-        "containerPath" : "container:iOSAppClient.xcodeproj",
+        "containerPath" : "container:HelloWorldiOSClientApp.xcodeproj",
         "identifier" : "3A133CF82B1F28890008DD5E",
-        "name" : "iOSAppClientUITests"
+        "name" : "HelloWorldiOSClientAppUITests"
       }
     }
   ],
