+++ Examples/HelloWorldiOSClientAppExample/HelloWorldiOSClientApp/Assets.xcassets/AccentColor.colorset/Contents.json
