+++ Examples/HelloWorldiOSClientAppExample/HelloWorldiOSClientApp/Assets.xcassets/AppIcon.appiconset/Contents.json
