+++ Examples/HelloWorldiOSClientAppExample/HelloWorldiOSClientApp/Assets.xcassets/Contents.json
