+++ Examples/HelloWorldiOSClientAppExample/HelloWorldiOSClientApp/Preview Content/Preview Assets.xcassets/Contents.json
